+++ src/typeguard/_transformer.py
@@ -376,6 +376,16 @@ class AnnotationTransformer(NodeTransformer):
         self.generic_visit(node)
 
         if isinstance(node.op, BitOr):
+            # If either branch of the BinOp has been transformed to `None`
+            # then the `ast.generic_visit` will eliminate that branch completely.
+            # If this happens, treat the BinOp as just the other branch.
+            if not hasattr(node, "left") and not hasattr(node, "right"):
+                return None
+            if not hasattr(node, "left"):
+                return node.right
+            if not hasattr(node, "right"):
+                return node.left
+
             # Return Any if either side is Any
             if self._memo.name_matches(node.left, *anytype_names):
                 return node.left